fix: use tenant path instead of tenant id for map index
The frontend needs to verify the dynamic segment in the route is a valid tenant path for the current
user. Therefore it make more sense for the map index to be the tenant path, not the tenant id.
